# Conflicts:
#	myopia-business/business-api/management-api/src/main/java/com/wupol/myopia/business/api/management/controller/StatManagementController.java

+++ myopia-business/business-core/screening-flow-core/src/main/java/com/wupol/myopia/business/core/screening/flow/util/ScreeningResultUtil.java
@@ -1524,7 +1524,7 @@ public class ScreeningResultUtil {
     private RefractoryResultItems.Item packageSpnItem(BigDecimal spn) {
         RefractoryResultItems.Item spnItems = new RefractoryResultItems.Item();
         spnItems.setVision(spn);
-        spnItems.setTypeName(spn.abs().multiply(new BigDecimal("100")) + "度");
+        spnItems.setTypeName(spn.abs().multiply(new BigDecimal("100")).intValue() + "度");
         return spnItems;
     }
 }
